# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: BlueBird Task API
  version: 0.1.0
  description: >
    蓝鸟任务系统 **目标态** API 契约（单一事实源，针对重构后的新系统；**非现有系统**——现有系统为 `/task/record/*` 裸 `Page`/裸实体、成功码由前端按「缺失即 200」解释，见 `docs/frontend-baseline/fixtures/legacy-api/`）。
    统一前缀 `/api/v1`；统一响应 `ApiResult{code,message,data,traceId}`，`code==0` 为成功、**HTTP 恒 200**（业务码表达结果，见 `02 §1.3`）。
    错误码区间（枚举权威源 `02 §1.4`）：`1xxxx=common`、`2xxxx=identity`、`3xxxx=task`、`4xxxx=file`、`5xxxx=预留`、`6xxxx=org/同步`。
    时间统一 ISO8601 + `+08:00`。详见 `Task/02后端模块详细设计.md`。
    约定：用户/个人数据响应统一 `Cache-Control: no-store`；秘密一律不入 URL（token 走 `Authorization` 头，密码/密钥走请求体；query 仅非敏感参数）。
servers:
  - url: /api/v1
    description: 同源（单制品内嵌前端，ADR-009）
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [task]
      summary: 任务列表（六大视图）
      parameters:
        - { name: scope, in: query, required: true, schema: { $ref: '#/components/schemas/Scope' } }
        - { name: keyword, in: query, schema: { type: string }, description: 匹配 title/content（大小写不敏感子串） }
        - { name: subordinate, in: query, schema: { type: boolean, default: false }, description: 仅 scope=all 生效；主管筛选其（含递归子部门）下属相关任务，只读（ADR-012） }
        - { name: date, in: query, schema: { type: string, format: date } }
        - { name: categoryId, in: query, schema: { type: integer, format: int64 }, description: 按分类过滤（含子树，ADR-015 共享分类） }
        - { name: menuId, in: query, schema: { type: integer, format: int64 }, description: 按自定义栏条目过滤（仅本人栏） }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200': { $ref: '#/components/responses/TaskPage' }
    post:
      tags: [task]
      summary: 新建主任务
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TaskCreateReq' } } }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
...
  /tasks/calendar:
    get:
      tags: [task]
      summary: 日历区间（有任务日期/实例）
      parameters:
        - { name: start, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: end, in: query, required: true, schema: { type: string, format: date-time } }
      responses:
        '200': { $ref: '#/components/responses/TaskList' }
  /tasks/subtasks:
    get:
      tags: [task]
      summary: 子任务列表
      parameters: [{ name: parentId, in: query, required: true, schema: { type: integer, format: int64 } }]
      responses:
        '200': { $ref: '#/components/responses/TaskList' }
  /tasks/{id}:
    get:
      tags: [task]
      summary: 任务详情
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf: [{ $ref: '#/components/schemas/ApiResult' }, { properties: { data: { $ref: '#/components/schemas/TaskDetailVO' } } }]
    put:
      tags: [task]
      summary: 更新任务（须携带 `version`）
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TaskUpdateReq' } } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [task]
      summary: 软删除（不可逆：清空关联数据；见 R5）
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /tasks/{id}/complete:
    post:
      tags: [task]
      summary: 完成任务（周期任务可传 dueAt 指定实例，见 R2；更新须携带 version）
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompleteReq'
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /tasks/{id}/uncomplete:
    post:
      tags: [task]
      summary: 取消完成（周期任务可传 dueAt；更新须携带 version）
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCompleteReq'
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /tasks/{id}/collect:
    post:
      tags: [task]
      summary: 收藏
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [task]
      summary: 取消收藏
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResult' }
...
    TaskList:
      description: 任务列表
      content:
        application/json:
          schema:
            allOf: [{ $ref: '#/components/schemas/ApiResult' }, { properties: { data: { $ref: '#/components/schemas/TaskListData' } } }]
    TaskPage:
      description: 任务分页
      content:
        application/json:
          schema:
            allOf: [{ $ref: '#/components/schemas/ApiResult' }, { properties: { data: { $ref: '#/components/schemas/TaskListData' } } }]
...
  schemas:
    ApiResult:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { nullable: true }
        traceId: { type: string, nullable: true }
...
    Scope:
      type: string
      enum: [day, week, assigned, joined, collect, all]
    Priority:
      type: string
      description: 任务优先级（4 值，与 01 §8.1 / 02 DDL 对齐）。
      enum: [LOW, MEDIUM, HIGH, URGENT]
    TaskStatus:
      type: string
      enum: [ACTIVE, DISABLED, ARCHIVED]
...
    CycleRule:
      type: object
      description: 周期规则（TEXT 存 JSON，计算式展开；02 §4.3）
      required: [freq, interval, dtstart, tz]
      properties:
        freq: { type: string, enum: [DAILY, WEEKLY, MONTHLY] }
        interval: { type: integer, minimum: 1, default: 1 }
        dtstart: { type: string, format: date-time }
        byDay: { type: array, items: { type: string, enum: [MO, TU, WE, TH, FR, SA, SU] }, nullable: true, description: 仅 WEEKLY 生效 }
        count: { type: integer, nullable: true, description: 与 until 互斥 }
        until: { type: string, format: date, nullable: true, description: 最后发生日期（含）；与 count 互斥。日期粒度（截至该日当地时区） }
        tz: { type: string, default: Asia/Shanghai }
    UserBrief:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
...
    TagVO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
    CategoryBrief:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
    FileVO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        fileName: { type: string }
        size: { type: integer, format: int64 }
    TaskVO:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        content: { type: string }
        status: { $ref: '#/components/schemas/TaskStatus' }
        completed: { type: boolean }
        dueAt: { type: string, format: date-time, nullable: true }
        remindAt: { type: string, format: date-time, nullable: true }
        priority: { $ref: '#/components/schemas/Priority' }
        cycleRule: { allOf: [{ $ref: '#/components/schemas/CycleRule' }], nullable: true }
        cycleLastCompleted: { type: string, format: date-time, nullable: true }
        owner: { $ref: '#/components/schemas/UserBrief' }
        assignees: { type: array, items: { $ref: '#/components/schemas/UserBrief' } }
        ccUsers: { type: array, items: { $ref: '#/components/schemas/UserBrief' } }
        participantIds: { type: array, items: { type: integer, format: int64 } }
        version: { type: integer, format: int64, description: 乐观锁版本号（更新须回传；失配 code=10006，ADR-013） }
        category: { allOf: [{ $ref: '#/components/schemas/CategoryBrief' }], nullable: true }
        tags: { type: array, items: { $ref: '#/components/schemas/TagVO' } }
        subtaskCompleted: { type: integer }
        subtaskTotal: { type: integer }
        files: { type: array, items: { $ref: '#/components/schemas/FileVO' } }
        createdAt: { type: string, format: date-time }
    TaskDetailVO:
      allOf:
        - $ref: '#/components/schemas/TaskVO'
        - type: object
          properties:
            subtasks: { type: array, items: { $ref: '#/components/schemas/TaskVO' } }
    TaskCreateReq:
      type: object
      required: [title]
      properties:
        title: { type: string }
        content: { type: string }
        version: { type: integer, format: int64, description: 更新（PUT）时必填乐观锁版本；新建省略（ADR-013） }
        dueAt: { type: string, format: date-time, nullable: true }
        remindAt: { type: string, format: date-time, nullable: true }
        priority: { $ref: '#/components/schemas/Priority' }
        cycleRule: { allOf: [{ $ref: '#/components/schemas/CycleRule' }], nullable: true }
        categoryId: { type: integer, format: int64, nullable: true }
        assigneeIds: { type: array, items: { type: integer, format: int64 } }
        ccIds: { type: array, items: { type: integer, format: int64 } }
        tagIds: { type: array, items: { type: integer, format: int64 } }
        fileIds: { type: array, items: { type: integer, format: int64 } }
    TaskListData:
      type: object
      properties:
        list: { type: array, items: { $ref: '#/components/schemas/TaskVO' } }
        total: { type: integer }
        page: { type: integer }
        size: { type: integer }
...
    TaskUpdateReq:
      description: 更新任务（在 TaskCreateReq 基础上 `version` 必填，ADR-013）
      allOf:
        - $ref: '#/components/schemas/TaskCreateReq'
        - type: object
          required: [version]
          properties:
            version: { type: integer, format: int64 }
    TaskCompleteReq:
      type: object
      description: 完成/取消完成请求体（ADR-013：更新须携带 `version`；周期任务需带 `dueAt` 指定实例，见 02 §4.3 R2）
      required: [version]
      properties:
        version: { type: integer, format: int64, description: 乐观锁版本号（失配 code=10006） }
        dueAt: { type: string, format: date-time, nullable: true, description: 周期任务指定实例时间；非周期可空 }
...